Remove ui.js and init.js
