+++ js/app.js
@@ -1,3 +1,5 @@
+/* global Notification, SearchProxy */
+
 /**
  * @author Christoph Wurst <christoph@winzerhof-wurst.at>
  *
@@ -22,7 +24,9 @@ define(function(require) {
 
 	var $ = require('jquery');
 	var Backbone = require('backbone');
+	var Handlebars = require('handlebars');
 	var Marionette = require('marionette');
+	var OC = require('OC');
 	var AppView = require('views/app');
 	var Radio = require('radio');
 	var Router = require('router');
@@ -37,7 +41,36 @@ define(function(require) {
 	require('service/messageservice');
 	require('notification');
 
-	var Mail = Marionette.Application.extend();
+	// Set marionette defaults
+	Marionette.TemplateCache.prototype.compileTemplate = function(rawTemplate) {
+		return Handlebars.compile(rawTemplate);
+	};
+	Marionette.ItemView.prototype.modelEvents = {change: 'render'};
+	Marionette.CompositeView.prototype.modelEvents = {change: 'render'};
+
+	var Mail = Marionette.Application.extend({
+		registerProtocolHandler: function() {
+			if (window.navigator.registerProtocolHandler) {
+				var url = window.location.protocol + '//' +
+					window.location.host +
+					OC.generateUrl('apps/mail/compose?uri=%s');
+				try {
+					window.navigator
+						.registerProtocolHandler('mailto', url, 'ownCloud Mail');
+				} catch (e) {
+				}
+			}
+		},
+		requestNotificationPermissions: function() {
+			// request permissions
+			if (typeof Notification !== 'undefined') {
+				Notification.requestPermission();
+			}
+		},
+		setUpSearch: function() {
+			SearchProxy.setFilter(require('search').filter);
+		}
+	});
 
 	Mail = new Mail();
 
@@ -45,6 +78,10 @@ define(function(require) {
 		this.view = new AppView();
 
 		Radio.ui.trigger('content:loading');
+
+		this.registerProtocolHandler();
+		this.requestNotificationPermissions();
+		this.setUpSearch();
 
 		$.when(AccountController.loadAccounts()).done(function(accounts) {
 			$('#app-navigation').removeClass('icon-loading');
@@ -57,7 +94,40 @@ define(function(require) {
 			});
 			Backbone.history.start();
 		});
+
+		/**
+		 * Detects pasted text by browser plugins, and other software.
+		 * Check for changes in message bodies every second.
+		 */
+		setInterval((function() {
+			// Begin the loop.
+			return function() {
+
+				// Define which elements hold the message body.
+				var MessageBody = $('.message-body');
+
+				/**
+				 * If the message body is displayed and has content:
+				 * Prepare the message body content for processing.
+				 * If there is new message body content to process:
+				 * Resize the text area.
+				 * Toggle the send button, based on whether the message is ready or not.
+				 * Prepare the new message body content for future processing.
+				 */
+				if (MessageBody.val()) {
+					var OldMessageBody = MessageBody.val();
+					var NewMessageBody = MessageBody.val();
+					if (NewMessageBody !== OldMessageBody) {
+						MessageBody.trigger('autosize.resize');
+						OldMessageBody = NewMessageBody;
+					}
+				}
+			};
+		})(), 1000);
 	});
 
+	_.delay(function() {
+		Mail.start();
+	});
 	return Mail;
 });
