+++ js/require_config.js
@@ -41,7 +41,6 @@
 
 	require([
 		'app',
-		'init',
 		'notification'
 	]);
 })();
