--- js/ui.js
@@ -1,116 +1,0 @@
-/**
- * ownCloud - Mail
- *
- * This file is licensed under the Affero General Public License version 3 or
- * later. See the COPYING file.
- *
- * @author Christoph Wurst <christoph@winzerhof-wurst.at>
- * @copyright Christoph Wurst 2016
- */
-
-define(function(require) {
-	'use strict';
-
-	var $ = require('jquery');
-	var OC = require('OC');
-	var Radio = require('radio');
-
-	require('views/helper');
-
-	var composerVisible = false;
-
-	/**
-	 * @param {number} messageId
-	 * @param {number} attachmentId
-	 * @returns {undefined}
-	 */
-	function saveAttachment(messageId, attachmentId) {
-		OC.dialogs.filepicker(
-			t('mail', 'Choose a folder to store the attachment in'),
-			function(path) {
-				// Loading feedback
-				var saveToFilesBtnSelector = '.attachment-save-to-cloud';
-				if (typeof attachmentId !== 'undefined') {
-					saveToFilesBtnSelector = 'li[data-attachment-id="' +
-						attachmentId + '"] ' + saveToFilesBtnSelector;
-				}
-				$(saveToFilesBtnSelector)
-					.removeClass('icon-folder')
-					.addClass('icon-loading-small')
-					.prop('disabled', true);
-
-				$.ajax(
-					OC.generateUrl(
-						'apps/mail/accounts/{accountId}/' +
-						'folders/{folderId}/messages/{messageId}/' +
-						'attachment/{attachmentId}',
-						{
-							accountId: require('state').currentAccount.get('accountId'),
-							folderId: require('state').currentFolder.get('id'),
-							messageId: messageId,
-							attachmentId: attachmentId
-						}), {
-					data: {
-						targetPath: path
-					},
-					type: 'POST',
-					success: function() {
-						if (typeof attachmentId === 'undefined') {
-							Radio.ui.trigger('error:show', t('mail', 'Attachments saved to Files.'));
-						} else {
-							Radio.ui.trigger('error:show', t('mail', 'Attachment saved to Files.'));
-						}
-					},
-					error: function() {
-						if (typeof attachmentId === 'undefined') {
-							Radio.ui.trigger('error:show', t('mail', 'Error while saving attachments to Files.'));
-						} else {
-							Radio.ui.trigger('error:show', t('mail', 'Error while saving attachment to Files.'));
-						}
-					},
-					complete: function() {
-						// Remove loading feedback again
-						$('.attachment-save-to-cloud')
-							.removeClass('icon-loading-small')
-							.addClass('icon-folder')
-							.prop('disabled', false);
-					}
-				});
-			},
-			false,
-			'httpd/unix-directory',
-			true
-			);
-	}
-
-	/**
-	 * @returns {undefined}
-	 */
-	function openForwardComposer() {
-		var header = '\n\n\n\n-------- ' +
-			t('mail', 'Forwarded message') +
-			' --------\n';
-
-		// TODO: find a better way to get the current message body
-		var data = {
-			subject: 'Fwd: ' + require('state').currentMessageSubject,
-			body: header + require('state').currentMessageBody.replace(/<br \/>/g, '\n')
-		};
-
-		if (require('state').currentAccount.get('accountId') !== -1) {
-			data.accountId = require('state').currentAccount.get('accountId');
-		}
-
-		Radio.ui.trigger('composer:show', data);
-	}
-
-	function isComposerVisible() {
-		return composerVisible;
-	}
-
-	return {
-		saveAttachment: saveAttachment,
-		openForwardComposer: openForwardComposer,
-		isComposerVisible: isComposerVisible
-	};
-});
